add logic for the rest of the tests and structured the project

+++ DataLayer/DTOs/OrderDetailDTO.cs
@@ -3,7 +3,7 @@ namespace DataLayer
     public class OrderDetailDTO
     {
         public int Quantity { get; set; }
-        public decimal UnitPrice { get; set; }
+        public int UnitPrice { get; set; }
         public ProductDetailDTO Product { get; set; }
     }
 }
